+++ package.json
@@ -19,7 +19,7 @@
     "babel-preset-flow": "7.0.0-alpha.18",
     "babel-preset-stage-0": "7.0.0-alpha.18",
     "babel-register": "7.0.0-alpha.18",
-    "babylon": "7.0.0-beta.22",
+    "babylon": "7.0.0-beta.25",
     "browserify": "^13.1.1",
     "bundle-collapser": "^1.2.1",
     "chai": "^4.1.0",
