+++ packages/babel-generator/package.json
@@ -20,6 +20,6 @@
   },
   "devDependencies": {
     "babel-helper-fixtures": "7.0.0-beta.1",
-    "babylon": "^7.0.0-beta.22"
+    "babylon": "^7.0.0-beta.25"
   }
 }
