+++ packages/babel-generator/src/generators/index.js
@@ -8,5 +8,4 @@ export * from "./types";
 export * from "./flow";
 export * from "./base";
 export * from "./jsx";
-export * from "./tsFlowCommon";
 export * from "./typescript";
