+++ packages/babel-template/package.json
@@ -10,7 +10,7 @@
   "dependencies": {
     "babel-traverse": "7.0.0-beta.1",
     "babel-types": "7.0.0-beta.1",
-    "babylon": "7.0.0-beta.22",
+    "babylon": "7.0.0-beta.25",
     "lodash": "^4.2.0"
   }
 }
