+++ packages/babel-types/src/definitions/experimental.js
@@ -1,4 +1,10 @@
-import defineType, { assertNodeType } from "./index";
+import defineType, {
+  assertEach,
+  assertNodeType,
+  assertValueType,
+  chain,
+} from "./index";
+import { classMethodOrPropertyCommon } from "./es2015";
 
 defineType("AwaitExpression", {
   builder: ["argument"],
@@ -16,6 +22,34 @@ defineType("BindExpression", {
   aliases: ["Expression"],
   fields: {
     // todo
+  },
+});
+
+defineType("ClassProperty", {
+  visitor: ["key", "value", "typeAnnotation", "decorators"],
+  builder: ["key", "value", "typeAnnotation", "decorators", "computed"],
+  aliases: ["Property"],
+  fields: {
+    ...classMethodOrPropertyCommon,
+    value: {
+      validate: assertNodeType("Expression"),
+      optional: true,
+    },
+    typeAnnotation: {
+      validate: assertNodeType("TypeAnnotation", "TSTypeAnnotation", "Noop"),
+      optional: true,
+    },
+    decorators: {
+      validate: chain(
+        assertValueType("array"),
+        assertEach(assertNodeType("Decorator")),
+      ),
+      optional: true,
+    },
+    readonly: {
+      validate: assertValueType("boolean"),
+      optional: true,
+    },
   },
 });
