+++ packages/babel-types/src/definitions/init.js
@@ -5,5 +5,4 @@ import "./flow";
 import "./jsx";
 import "./misc";
 import "./experimental";
-import "./tsFlowCommon";
 import "./typescript";
